# public/.well-known/zakai-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zakai Rights Protocol API",
    "version": "2026-08-01",
    "description": "Public API for consumer rights discovery. Catalog access requires no PII.",
    "contact": { "name": "Zakai Protocol", "url": "https://zakai-3uxj.vercel.app" }
  },
  "servers": [{ "url": "/api", "description": "Same origin as the Zakai deployment" }],
  "paths": {
...
    "/fairness/scores": {
      "get": {
        "operationId": "listFairnessScores",
        "summary": "De-identified provider win rates (0–100) from StrategyOutcome",
        "tags": ["Fairness"],
        "parameters": [
          {
            "name": "market",
            "in": "query",
            "schema": { "type": "string", "default": "IL", "pattern": "^[A-Z]{2}$" }
          }
        ],
        "responses": {
          "200": {
            "description": "Scores (empty until MIN_SAMPLE observations per provider)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FairnessScoresResponse" }
              }
            }
          },
          "429": { "description": "Rate limited" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FairnessScoresResponse": {
        "type": "object",
        "required": ["market", "min_observations", "providers", "disclaimer"],
        "properties": {
          "market": { "type": "string" },
          "min_observations": { "type": "integer" },
          "disclaimer": { "type": "string" },
          "providers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FairnessProviderScore" }
          },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "FairnessProviderScore": {
        "type": "object",
        "properties": {
          "provider": { "type": "string" },
          "fairnessScore": { "type": "integer", "minimum": 0, "maximum": 100 },
          "observations": { "type": "integer" },
          "wins": { "type": "integer" },
          "methodology": { "type": "string", "enum": ["strategy_outcome_win_rate"] }
        }
      }
    }
  }
}